#192 save and load the tick titles
Instead of the name of the layer, the tick title plus a certain suffix
is saved in the config file. When a tick title for a layer is needed,
then there are two cases. If the layer name contains the suffix, then
the suffix has to be removed and the layer name is the tick title.
Otherwise a shortened tick title has to be created has this was
previously the case.
The suffix is needed as the name of the layer is overwritten, when the
layer loads its map. Therefore the shortened title has to be created
again.