#198 introduced the print mode
The printing does not work because if only a single layer is loaded, the
needed retrieval complete event will never be fired, which the
headlessmapProvider needs to work.
If this event is fired after a complete retrieval and if if the layer is
locked, then the sliding functionality will not work anymore.
Therefore the event is fired, if one layer was retrieved and when the
layer is in the print mode.