#16 added a new rectangle which draws the border
In the previous attempts it was tried to draw the border of the
printingRectangle directly. This lead to a misbehavior when the
rectangle was moved as it grew. Thus another rectangle was added which
gets the same borders as the printingRectangle. Due to this additional
rectangle the wished border can be drawn.